feat: 아이템 삭제 시, 해당 아이템의 스페이스에 대한 랜덤 아이템이 Redis에 존재하는 경우에 대한 삭제 거부 로직 추가

+++ src/main/java/com/example/rqs/api/item/ItemController.java
@@ -3,6 +3,7 @@ package com.example.rqs.api.item;
 import com.example.rqs.api.cache.randomItem.RandomItemCache;
 import com.example.rqs.api.cache.randomItem.RandomItemCacheService;
 import com.example.rqs.api.jwt.MemberDetails;
+import com.example.rqs.core.common.exception.BadRequestException;
 import com.example.rqs.core.item.service.ItemService;
 import com.example.rqs.core.item.service.dtos.*;
 
@@ -123,8 +124,11 @@ public class ItemController {
     @DeleteMapping("")
     public DeleteResponse deleteItem(
             @AuthenticationPrincipal MemberDetails memberDetails,
+            @RequestParam("spaceId") Long spaceId,
             @RequestParam("itemId") Long itemId
     ) {
+        boolean exist = randomItemCacheService.existCacheByKeyPattern(spaceId + "_*");
+        if (exist) throw new BadRequestException("현재 아이템을 삭제할 수 없습니다. 5분 뒤에 다시 시도해보세요.");
         DeleteItem deleteItem = DeleteItem.of(memberDetails.getMember(), itemId);
         itemService.deleteItem(deleteItem);
         return DeleteResponse.of(itemId, true);
